+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Please refer to the [NEWS](NEWS.md) for a list of changes which have an affect o
 
 ### Bots
 #### Collectors
+- `intelmq.bots.collectors.http.collector_http`: Add Chunking parameters to handle big files (PR#2684 by Sebastian Wagner).
 
 #### Parsers
 - `intelmq.bots.parsers.microsoft.parser_ctip`: Use the new field `serverity` instead of `extra.severity` (PR#2662 by Sebastian Wagner).